Disable MDL for next-action checkboxes
d1d071462a0a1520d94a029886093b558b24267d causes uGTD to dynamically
regenerate the next action list when you create a new action.
Unfortunately, to get the new checkboxes to look right, it has to
MDL.upgrade them, and this clobbers the onclick handler that Ur/Web
installs.  Disable MDL on the checkboxes to preserve functionality (at
the cost of a bit of ugliness).